+++ src/equicordplugins/moreStickers/stickers.ts
@@ -8,7 +8,6 @@ import * as DataStore from "@api/DataStore";
 
 import { removeRecentStickerByPackId } from "./components";
 import { DynamicStickerPackMeta, StickerPack, StickerPackMeta } from "./types";
-import { corsFetch } from "./utils";
 
 const PACKS_KEY = "MoreStickers:Packs";
 
@@ -82,8 +81,9 @@ export async function deleteStickerPack(id: string, packsKey: string = PACKS_KEY
 // ---------------------------- Dynamic Packs ----------------------------
 
 export async function getDynamicStickerPack(dspm: DynamicStickerPackMeta): Promise<StickerPack | null> {
-    const dsp = await corsFetch(dspm.dynamic.refreshUrl, {
+    const dsp = await fetch(dspm.dynamic.refreshUrl, {
         headers: dspm.dynamic.authHeaders,
+        redirect: "error",
     });
     if (!dsp.ok) return null;
     return await dsp.json();